# btree/tablebtree_move_to: micro-optimizations