@jussisaurio 2025-04-26
explicitly removing the integer division for the binary search halfpoint calculation seems to surprisingly make a small additional difference: https://github.com/tursodatabase/limbo/pull/1408/commits/0ebc145345de198c701093dff66bf1b8ff93b084